requirements.txt
wheel
setuptools
cython
tqdm

numpy
pandas
Pillow
gdal
geojson
rasterio